requirements.txt
streamlit
torch
torchvision
opencv-python
pillow
fpdf
gdown